TINY-7478: Normalized border styles with dialog, extracted advanced dialog.
